feat(brain): first-class hal0-brain module + /api/brain/chat primary route
Extract the hal0-brain steward chat into a first-class package
`src/hal0/brain/` that consumes the shared P2 tool-loop engine
(`hal0.toolloop.engine`) and has ZERO Hermes/board import dependency
(SPEC §G / R4). The board's operator-board client and the platform
self-API are reached only through `app.state` handles injected at
runtime — nothing in `src/hal0/brain/` imports them.

- New PRIMARY route `POST /api/brain/chat` (api/routes/brain.py),
  mounted pre-dashboard, ADMIN-gated (deny-by-default) via a new
  `/api/brain` rule in security/exposure.py.
- `api/routes/board_chat.py` becomes a thin alias: it rebinds itself to
  the `hal0.brain.chat` module object in `sys.modules`, so every existing
  import path (`run_board_chat`, private `_foo` helpers,
  `monkeypatch.setattr(bc, ...)`) resolves to one shared namespace — no
  logic duplication, no drift. `/api/board/chat` and its KB-1 auth gating
  are unchanged.
- New tests/brain: primary-route SSE, thin-alias identity, and an
  AST-based no-Hermes/board-import invariant guard.

Verify (PYTHONPATH=src): ruff check + format clean; exposure CI green
(both /api/board/chat and /api/brain/chat classify ADMIN); board_chat
suite 96 passed via the alias; brain suite 3 passed; check_sunset scar
202 (unchanged).
